move to doubles?

+++ src/main/java/nl/jasper/flappy/GameObject.java
@@ -3,6 +3,7 @@ package nl.jasper.flappy;
 import nl.jasper.flappy.render.Mesh;
 import nl.jasper.flappy.render.Shader;
 import nl.jasper.flappy.render.Texture;
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 public abstract class GameObject {
@@ -10,9 +11,9 @@ public abstract class GameObject {
     private Mesh mesh;
     private Texture texture;
 
-    public Vector3f position = new Vector3f();
+    public Vector3d position = new Vector3d();
     public float rotation = 0;
-    public Vector3f scale = new Vector3f(1);
+    public Vector3d scale = new Vector3d(1);
 
 
     public GameObject(Mesh mesh, Texture texture) {
